fix(sw): defensive cleanup + cast hygiene from triple SW review
Result of Gemini + Copilot + own deep-read audit (~21k LOC across
baremetal/, linux/, driver/, lib/, app/, cnn_eval/, lprnet_eval/).
Code quality was rated 9/10 overall; cgra.h, cgra_driver.c, lpr_golden.c,
and demo_mnist_hdmi_bm.c specifically praised. These are the few
genuine, non-cosmetic findings.

linux/lpr_eval.c
  - ftell() return now validated (sz < 0 caught before malloc)
  - malloc() now NULL-checked before fread() (prevents SEGV on OOM)
  - fclose() now called on the short-read error path (was leaked)
  - CGRA-path cleanup: free(int16_blob) + close(memfd) added at exit
    (pool2_va/staging_va mmaps left for kernel reclaim — they're
    MAP_FIXED scratch regions, munmap on exit is redundant)

linux/lpr_hdmi_demo.c
  - Same pattern: ftell guard, malloc NULL-check, fread return checked,
    memfd open() failure now caught (was unchecked, would mmap on -1)

lib/cgra_tiler.c
  - calloc((size_t)total_rows * total_cols, ...) had cast on only one
    operand; added explicit (size_t) on the other for symmetry. C's
    implicit conversion already widens correctly on 32-bit ARM where
    size_t == unsigned int, but the explicit form is clearer and
    overflow-safe on 64-bit hosts.

Findings deliberately skipped:
  - cgra_driver.c strtol → strtoul: sysfs map sizes never set high bit,
    safe as-is.
  - cgra_driver.c uninitialized dev->fd: false positive — public entry
    point cgra_init() does memset(dev, 0, sizeof(*dev)); dev->fd = -1
    before any open() call (line 111).
  - demo_mnist_hdmi_bm.c (int) casts on size_t loop bounds: deliberate
    for compiler-warning suppression, project pattern.
  - Magic numbers in cnn_conv_bm.c (INPUT_SCALE=256, etc.): tied to
    specific quantized MNIST model, would migrate together if retrained.
  - golden_*/lpr_* prefix mix: stable API across many files, renaming
    would churn 50+ call sites for cosmetic gain.
  - "Header guards as #pragma once": both forms work, project mix is
    consistent within each subdir.

+++ 07_sw/lib/cgra_tiler.c
@@ -60,7 +60,7 @@ int im2col_4x4_tile(const int32_t *data_im,
     int total_cols = out_h * out_w;
 
     /* Full Im2Col into temporary buffer */
-    int32_t *col = (int32_t *)calloc((size_t)total_rows * total_cols,
+    int32_t *col = (int32_t *)calloc((size_t)total_rows * (size_t)total_cols,
                                       sizeof(int32_t));
     if (!col) return -1;
 
@@ -68,7 +68,7 @@ int im2col_4x4_tile(const int32_t *data_im,
 
     /* Slice into 4-row tiles; last tile zero-padded if needed */
     int nt = (total_rows + 3) / 4;
-    int32_t *tile_buf = (int32_t *)calloc((size_t)nt * 4 * total_cols,
+    int32_t *tile_buf = (int32_t *)calloc((size_t)nt * 4 * (size_t)total_cols,
                                            sizeof(int32_t));
     if (!tile_buf) { free(col); return -1; }
